requirements.txt
absl-py==0.12.0
anyio==3.3.4
argon2-cffi==20.1.0
asgiref==3.4.1
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
cloudpickle==2.0.0
contextlib2==21.6.0
contextvars==2.4
cryptography==2.1.4
cycler==0.10.0
dask==2021.3.0
dask-glm==0.2.0
dask-ml==1.9.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
distributed==2021.3.0
entrypoints==0.3
fastapi==0.70.0
flatbuffers==1.12
fsspec==2021.10.1
gast==0.4.0
gensim==4.1.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
HeapDict==1.0.1
idna==3.3
immutables==0.16
importlib-metadata==4.0.1
iniconfig==1.1.1
ipykernel==5.1.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.0
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.3.4
mistune==0.8.4
# Editable install with no version control (mlcore==0.1.0)
-e /recsysapp/mlcore
msgpack==1.0.2
multipledispatch==0.6.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
node2vec==0.4.3
notebook==6.3.0
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==1.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pycrypto==2.6.1
pydantic==1.8.2
Pygments==2.9.0
PyGObject==3.26.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.5
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-dotenv==0.19.1
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.15.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.16.0
tblib==1.7.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
terminado==0.9.5
testpath==0.4.4
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zict==2.0.0
zipp==3.4.1
